# artifacts/api-server/src/routes/battles.ts
import { Router, type IRouter } from "express";
import { db } from "@workspace/db";
import {
  battlesTable,
  capturedMonstersTable,
  monsterSpeciesTable,
  playersTable,
} from "@workspace/db";
import { eq, and } from "drizzle-orm";
import { requireAuth } from "../middlewares/requireAuth.js";
import { formatSpecies } from "./monsters.js";
import { formatCapturedMonster } from "./collection.js";
import {
  StartBattleBody,
  StartBattleResponse,
  GetBattleParams,
  GetBattleResponse,
  PerformBattleActionParams,
  PerformBattleActionBody,
  PerformBattleActionResponse,
} from "@workspace/api-zod";
import {
  calculateDamage,
  calculateCaptureChance,
  getElementMultiplier,
  rollCritical,
  calculateWildStats,
  xpForNextLevel,
  getExplorerRankForLevel,
} from "../lib/gameEngine.js";
import type { BattleLogEntry } from "@workspace/db";

const router: IRouter = Router();

function formatBattle(
  b: typeof battlesTable.$inferSelect,
  species: typeof monsterSpeciesTable.$inferSelect,
  playerMonsterRow: {
    captured: typeof capturedMonstersTable.$inferSelect;
    species: typeof monsterSpeciesTable.$inferSelect;
  } | null,
) {
  return {
    id: b.id,
    playerId: b.playerId,
    status: b.status,
    turn: b.turn,
    wildMonster: {
      speciesId: b.wildSpeciesId,
      capturedId: null,
      species: formatSpecies(species),
      level: b.wildLevel,
      currentHp: b.wildCurrentHp,
      maxHp: b.wildMaxHp,
      attack: b.wildAttack,
      defense: b.wildDefense,
      speed: b.wildSpeed,
      shinyVariant: b.wildShinyVariant ?? null,
      statusEffect: b.wildStatusEffect ?? null,
    },
    playerMonster: playerMonsterRow
      ? {
          speciesId: playerMonsterRow.captured.speciesId,
          capturedId: playerMonsterRow.captured.id,
          species: formatSpecies(playerMonsterRow.species),
          level: playerMonsterRow.captured.level,
          currentHp: b.playerCurrentHp,
          maxHp: playerMonsterRow.captured.maxHp,
          attack: playerMonsterRow.captured.attack,
          defense: playerMonsterRow.captured.defense,
          speed: playerMonsterRow.captured.speed,
          shinyVariant: playerMonsterRow.captured.shinyVariant ?? null,
          statusEffect: b.playerStatusEffect ?? null,
        }
      : null,
    log: b.log as BattleLogEntry[],
    expReward: b.expReward ?? null,
    coinReward: b.coinReward ?? null,
    capturedMonsterId: b.capturedMonsterId ?? null,
    regionId: b.regionId,
    createdAt: b.createdAt.toISOString(),
    updatedAt: b.updatedAt.toISOString(),
  };
}

// POST /battles
router.post("/battles", requireAuth, async (req, res): Promise<void> => {
  const body = StartBattleBody.safeParse(req.body);
  if (!body.success) {
    res.status(400).json({ error: body.error.message });
    return;
  }
  if (req.playerId !== body.data.playerId) {
    res.status(403).json({ error: "Forbidden" });
    return;
  }

  const [species] = await db
    .select()
    .from(monsterSpeciesTable)
    .where(eq(monsterSpeciesTable.id, body.data.speciesId));
  if (!species) {
    res.status(404).json({ error: "Monster species not found" });
    return;
  }

  const [playerCaptured] = await db
    .select()
    .from(capturedMonstersTable)
    .where(
      and(
        eq(capturedMonstersTable.id, body.data.activeMonsterCapturedId),
        eq(capturedMonstersTable.playerId, body.data.playerId),
      ),
    );
  if (!playerCaptured) {
    res.status(404).json({ error: "Player monster not found" });
    return;
  }

  const wildStats = calculateWildStats(species, body.data.wildLevel);

  const [battle] = await db
    .insert(battlesTable)
    .values({
      playerId: body.data.playerId,
      status: "active",
      turn: 1,
      wildSpeciesId: species.id,
      wildLevel: body.data.wildLevel,
      wildCurrentHp: wildStats.hp,
      wildMaxHp: wildStats.hp,
      wildAttack: wildStats.attack,
      wildDefense: wildStats.defense,
      wildSpeed: wildStats.speed,
      wildShinyVariant: body.data.shinyVariant ?? null,
      playerCapturedId: playerCaptured.id,
      playerCurrentHp: playerCaptured.currentHp,
      regionId: body.data.regionId,
      log: [],
    })
    .returning();

  const [playerSpecies] = await db
    .select()
    .from(monsterSpeciesTable)
    .where(eq(monsterSpeciesTable.id, playerCaptured.speciesId));

  res.status(201).json(
    StartBattleResponse.parse(
      formatBattle(battle!, species, playerSpecies ? { captured: playerCaptured, species: playerSpecies } : null),
    ),
  );
});

// GET /battles/:battleId
router.get(
  "/battles/:battleId",
  requireAuth,
  async (req, res): Promise<void> => {
    const params = GetBattleParams.safeParse(req.params);
    if (!params.success) {
      res.status(400).json({ error: params.error.message });
      return;
    }
    const [battle] = await db
      .select()
      .from(battlesTable)
      .where(eq(battlesTable.id, params.data.battleId));
    if (!battle || battle.playerId !== req.playerId) {
      res.status(404).json({ error: "Battle not found" });
      return;
    }

    const [species] = await db
      .select()
      .from(monsterSpeciesTable)
      .where(eq(monsterSpeciesTable.id, battle.wildSpeciesId));
    const [playerCaptured] = await db
      .select()
      .from(capturedMonstersTable)
      .where(eq(capturedMonstersTable.id, battle.playerCapturedId));
    const [playerSpecies] = playerCaptured
      ? await db
          .select()
          .from(monsterSpeciesTable)
          .where(eq(monsterSpeciesTable.id, playerCaptured.speciesId))
      : [];

    res.json(
      GetBattleResponse.parse(
        formatBattle(
          battle,
          species!,
          playerCaptured && playerSpecies
            ? { captured: playerCaptured, species: playerSpecies }
            : null,
        ),
      ),
    );
  },
);

// POST /battles/:battleId/action
router.post(
  "/battles/:battleId/action",
  requireAuth,
  async (req, res): Promise<void> => {
    const params = PerformBattleActionParams.safeParse(req.params);
    if (!params.success) {
      res.status(400).json({ error: params.error.message });
      return;
    }
    const body = PerformBattleActionBody.safeParse(req.body);
    if (!body.success) {
      res.status(400).json({ error: body.error.message });
      return;
    }

    const [battle] = await db
      .select()
      .from(battlesTable)
      .where(eq(battlesTable.id, params.data.battleId));
    if (!battle || battle.playerId !== req.playerId) {
      res.status(404).json({ error: "Battle not found" });
      return;
    }
    if (battle.status !== "active") {
      res.status(400).json({ error: "Battle is already over" });
      return;
    }

    const [wildSpecies] = await db
      .select()
      .from(monsterSpeciesTable)
      .where(eq(monsterSpeciesTable.id, battle.wildSpeciesId));
    const [playerCaptured] = await db
      .select()
      .from(capturedMonstersTable)
      .where(eq(capturedMonstersTable.id, battle.playerCapturedId));
    const [playerSpecies] = playerCaptured
      ? await db
          .select()
          .from(monsterSpeciesTable)
          .where(eq(monsterSpeciesTable.id, playerCaptured.speciesId))
      : [];

    if (!wildSpecies || !playerCaptured || !playerSpecies) {
      res.status(500).json({ error: "Battle data corrupted" });
      return;
    }

    const log = (battle.log as BattleLogEntry[]) ?? [];
    let wildHp = battle.wildCurrentHp;
    let playerHp = battle.playerCurrentHp;
    let newStatus = "active";
    let capturedMonsterId: string | null = null;

    const action = body.data.action;
    const skills = (wildSpecies.skills as { name: string; type: string; element: string; power: number; accuracy: number }[]);
    const playerSkills = (playerSpecies.skills as typeof skills);

    if (action === "flee") {
      const fleeChance = playerCaptured.speed > battle.wildSpeed ? 0.9 : 0.5;
      if (Math.random() < fleeChance) {
        newStatus = "fled";
        log.push({
          turn: battle.turn,
          actor: "player",
          action: "flee",
          description: "You successfully fled from battle!",
          damageDealt: null,
          critical: false,
        });
      } else {
        log.push({
          turn: battle.turn,
          actor: "player",
          action: "flee",
          description: "Couldn't escape!",
          damageDealt: null,
          critical: false,
        });
      }
    } else if (action === "capture") {
      const orbType = body.data.orbType ?? "Basic";
      const success = calculateCaptureChance(
        orbType,
        wildHp,
        battle.wildMaxHp,
        wildSpecies.captureRate,
        battle.wildShinyVariant,
      );
      if (success) {
        // Create captured monster
        const personalities = ["Hardy", "Brave", "Calm", "Gentle", "Lax", "Bold", "Jolly", "Quirky", "Sassy", "Timid"];
        const personality = personalities[Math.floor(Math.random() * personalities.length)]!;
        const wildStats = calculateWildStats(wildSpecies, battle.wildLevel);
        const [captured] = await db
          .insert(capturedMonstersTable)
          .values({
            playerId: battle.playerId,
            speciesId: wildSpecies.id,
            level: battle.wildLevel,
            currentHp: wildStats.hp,
            maxHp: wildStats.hp,
            attack: wildStats.attack,
            defense: wildStats.defense,
            speed: wildStats.speed,
            shinyVariant: battle.wildShinyVariant ?? null,
            personality,
            inTeam: false,
          })
          .returning();
        capturedMonsterId = captured!.id;
        newStatus = "captured";
        // Update player captured count
        await db
          .update(playersTable)
          .set({ monstersCaptured: playerCaptured.level + 1 })
          .where(eq(playersTable.id, battle.playerId));
        log.push({
          turn: battle.turn,
          actor: "player",
          action: "capture",
          description: `You captured ${wildSpecies.name}!`,
          damageDealt: null,
          critical: false,
        });
      } else {
        log.push({
          turn: battle.turn,
          actor: "player",
          action: "capture",
          description: `${wildSpecies.name} broke free!`,
          damageDealt: null,
          critical: false,
        });
      }
    } else {
      // Attack or skill
      let skillPower = 40;
      let skillName = "Attack";
      let skillElement = playerSpecies.element;

      if (action === "attack") {
        const normalSkill = playerSkills.find((s) => s.type === "normal");
        if (normalSkill) {
          skillPower = normalSkill.power;
          skillName = normalSkill.name;
          skillElement = normalSkill.element;
        }
      } else if (action === "skill1") {
        const skill = playerSkills.find((s) => s.type === "skill1");
        if (skill) {
          skillPower = skill.power;
          skillName = skill.name;
          skillElement = skill.element;
        }
      } else if (action === "skill2") {
        const skill = playerSkills.find((s) => s.type === "skill2");
        if (skill) {
          skillPower = skill.power;
          skillName = skill.name;
          skillElement = skill.element;
        }
      } else if (action === "ultimate") {
        const skill = playerSkills.find((s) => s.type === "ultimate");
        if (skill) {
          skillPower = skill.power;
          skillName = skill.name;
          skillElement = skill.element;
        }
      }

      const critP = rollCritical();
      const multiplier = getElementMultiplier(skillElement, wildSpecies.element);
      const dmg = calculateDamage(
        playerCaptured.attack,
        battle.wildDefense,
        skillPower,
        multiplier,
        critP,
      );
      wildHp = Math.max(0, wildHp - dmg);

      let desc = `${playerCaptured.nickname ?? playerSpecies.name} used ${skillName} dealing ${dmg} damage!`;
      if (multiplier > 1) desc += " It's super effective!";
      if (multiplier < 1) desc += " It's not very effective...";
      if (critP) desc += " Critical hit!";

      log.push({
        turn: battle.turn,
        actor: "player",
        action,
        description: desc,
        damageDealt: dmg,
        critical: critP,
      });

      if (wildHp <= 0) {
        newStatus = "won";
        const expReward = Math.round(wildSpecies.baseHp * battle.wildLevel * 0.5);
        const coinReward = Math.round(10 + battle.wildLevel * 2);

        // Apply EXP to player monster
        const newExp = playerCaptured.experience + expReward;
        const newLevel = playerCaptured.level + Math.floor(newExp / xpForNextLevel(playerCaptured.level));
        await db
          .update(capturedMonstersTable)
          .set({
            experience: newExp % xpForNextLevel(playerCaptured.level),
            level: Math.min(100, newLevel),
          })
          .where(eq(capturedMonstersTable.id, playerCaptured.id));

        // Apply coins
        const [p] = await db
          .select({ coins: playersTable.coins, battlesWon: playersTable.battlesWon })
          .from(playersTable)
          .where(eq(playersTable.id, battle.playerId));
        await db
          .update(playersTable)
          .set({ coins: (p?.coins ?? 0) + coinReward, battlesWon: (p?.battlesWon ?? 0) + 1 })
          .where(eq(playersTable.id, battle.playerId));

        log.push({
          turn: battle.turn,
          actor: "player",
          action: "win",
          description: `You defeated ${wildSpecies.name}! +${expReward} EXP, +${coinReward} coins`,
          damageDealt: null,
          critical: false,
        });

        const [updatedBattle] = await db
          .update(battlesTable)
          .set({
            status: "won",
            wildCurrentHp: 0,
            log,
            expReward,
            coinReward,
          })
          .where(eq(battlesTable.id, battle.id))
          .returning();

        res.json(
          PerformBattleActionResponse.parse(
            formatBattle(
              updatedBattle!,
              wildSpecies,
              { captured: playerCaptured, species: playerSpecies },
            ),
          ),
        );
        return;
      }
    }

    // Wild monster attacks back if battle still active
    if (newStatus === "active") {
      const wildNormalSkill = skills.find((s) => s.type === "normal");
      const wildSkillPower = wildNormalSkill?.power ?? 40;
      const wildSkillName = wildNormalSkill?.name ?? "Attack";
      const wildSkillElement = wildNormalSkill?.element ?? wildSpecies.element;

      const wCrit = rollCritical();
      const wMult = getElementMultiplier(wildSkillElement, playerSpecies.element);
      const wDmg = calculateDamage(
        battle.wildAttack,
        playerCaptured.defense,
        wildSkillPower,
        wMult,
        wCrit,
      );
      playerHp = Math.max(0, playerHp - wDmg);

      log.push({
        turn: battle.turn,
        actor: "wild",
        action: wildSkillName,
        description: `Wild ${wildSpecies.name} used ${wildSkillName} dealing ${wDmg} damage!`,
        damageDealt: wDmg,
        critical: wCrit,
      });

      if (playerHp <= 0) {
        newStatus = "lost";
        const [p] = await db
          .select({ battlesLost: playersTable.battlesLost })
          .from(playersTable)
          .where(eq(playersTable.id, battle.playerId));
        await db
          .update(playersTable)
          .set({ battlesLost: (p?.battlesLost ?? 0) + 1 })
          .where(eq(playersTable.id, battle.playerId));
      }
    }

    const [updatedBattle] = await db
      .update(battlesTable)
      .set({
        status: newStatus,
        turn: battle.turn + 1,
        wildCurrentHp: wildHp,
        playerCurrentHp: playerHp,
        log,
        capturedMonsterId: capturedMonsterId ?? null,
      })
      .where(eq(battlesTable.id, battle.id))
      .returning();

    // Persist player monster HP change
    await db
      .update(capturedMonstersTable)
      .set({ currentHp: playerHp })
      .where(eq(capturedMonstersTable.id, playerCaptured.id));

    res.json(
      PerformBattleActionResponse.parse(
        formatBattle(
          updatedBattle!,
          wildSpecies,
          { captured: playerCaptured, species: playerSpecies },
        ),
      ),
    );
  },
);

export default router;