# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Monster Realms Online API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /battles:
    post:
      operationId: startBattle
      tags: [battles]
      summary: Start a wild monster battle
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartBattleInput"
      responses:
        "201":
          description: Battle created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Battle"
  /battles/{battleId}:
    get:
      operationId: getBattle
      tags: [battles]
      summary: Get current battle state
      security:
        - bearerAuth: []
      parameters:
        - name: battleId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Battle state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Battle"
        "404":
          description: Battle not found
  /battles/{battleId}/action:
    post:
      operationId: performBattleAction
      tags: [battles]
      summary: Perform a battle action (attack, skill, capture, flee)
      security:
        - bearerAuth: []
      parameters:
        - name: battleId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BattleActionInput"
      responses:
        "200":
          description: Updated battle state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Battle"

  # REGIONS
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MonsterSkill:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [normal, skill1, skill2, ultimate, passive]
        element:
          type: string
        power:
          type: number
        accuracy:
          type: number
        description:
          type: string
      required:
        - name
        - type
        - element
        - power
        - accuracy
        - description
    MonsterSpecies:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        element:
          type: string
          enum: [Fire, Water, Nature, Electric, Ice, Earth, Air, Light, Dark, Metal, Crystal, Void]
        rarity:
          type: string
          enum: [Common, Uncommon, Rare, Epic, Legendary, Mythic, Ancient, Celestial, Void]
        baseHp:
          type: number
        baseAttack:
          type: number
        baseDefense:
          type: number
        baseSpeed:
          type: number
        description:
          type: string
        lore:
          type: string
        captureRate:
          type: number
        weight:
          type: number
        height:
          type: number
        personality:
          type: string
        evolutionFromId:
          type: ["string", "null"]
        evolutionToId:
          type: ["string", "null"]
        evolutionLevel:
          type: ["number", "null"]
        skills:
          type: array
          items:
            $ref: "#/components/schemas/MonsterSkill"
        regionIds:
          type: array
          items:
            type: string
      required:
        - id
        - name
        - element
        - rarity
        - baseHp
        - baseAttack
        - baseDefense
        - baseSpeed
        - description
        - lore
        - captureRate
        - weight
        - height
        - personality
        - skills
        - regionIds

    # COLLECTION
...
    BattleMonsterState:
      type: object
      properties:
        speciesId:
          type: string
        capturedId:
          type: ["string", "null"]
        species:
          $ref: "#/components/schemas/MonsterSpecies"
        level:
          type: number
        currentHp:
          type: number
        maxHp:
          type: number
        attack:
          type: number
        defense:
          type: number
        speed:
          type: number
        shinyVariant:
          type: ["string", "null"]
        statusEffect:
          type: ["string", "null"]
      required:
        - speciesId
        - species
        - level
        - currentHp
        - maxHp
        - attack
        - defense
        - speed
    BattleLogEntry:
      type: object
      properties:
        turn:
          type: number
        actor:
          type: string
          enum: [player, wild]
        action:
          type: string
        description:
          type: string
        damageDealt:
          type: ["number", "null"]
        critical:
          type: boolean
      required:
        - turn
        - actor
        - action
        - description
        - critical
    Battle:
      type: object
      properties:
        id:
          type: string
        playerId:
          type: string
        status:
          type: string
          enum: [active, won, lost, fled, captured]
        turn:
          type: number
        wildMonster:
          $ref: "#/components/schemas/BattleMonsterState"
        playerMonster:
          $ref: "#/components/schemas/BattleMonsterState"
        log:
          type: array
          items:
            $ref: "#/components/schemas/BattleLogEntry"
        expReward:
          type: ["number", "null"]
        coinReward:
          type: ["number", "null"]
        capturedMonsterId:
          type: ["string", "null"]
        regionId:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - playerId
        - status
        - turn
        - wildMonster
        - playerMonster
        - log
        - regionId
        - createdAt
        - updatedAt
    StartBattleInput:
      type: object
      properties:
        playerId:
          type: string
        regionId:
          type: string
        speciesId:
          type: string
        wildLevel:
          type: number
        shinyVariant:
          type: ["string", "null"]
          enum: [Golden, Crystal, Shadow, Galaxy, Prismatic, null]
        activeMonsterCapturedId:
          type: string
      required:
        - playerId
        - regionId
        - speciesId
        - wildLevel
        - activeMonsterCapturedId
    BattleActionInput:
      type: object
      properties:
        action:
          type: string
          enum: [attack, skill1, skill2, ultimate, capture, flee]
        orbType:
          type: ["string", "null"]
          enum: [Basic, Explorer, Hunter, Elite, Master, Celestial, Infinity, null]
      required:
        - action

    # REGIONS
...